requirements.txt
Flask[async]==2.2.5
python-dotenv==1.0.0
asyncio~=3.4.3
aiohttp~=3.8.4
setuptools~=65.5.0
ghunt~=2.0.3
httpx
holehe